It performs the necessary function. Next step: write unit tests

+++ src/main/java/process/definitions/Division.java
@@ -1,7 +1,7 @@
 package process.definitions;
 
 public class Division implements ArithmeticWithBothX{
-    public int doOperation(int x1, int x2){
+    public double doOperation(double x1, double x2){
         return (x1/x2);
     }
 }
