+++ src/main/java/process/definitions/Square.java
@@ -1,7 +1,7 @@
 package process.definitions;
 
 public class Square implements ArithmeticWithX1Only{
-    public int doOperation(int x1){
+    public double doOperation(double x1){
         return (x1*x1);
     }
 }
