+++ src/main/java/process/definitions/SquareRoot.java
@@ -1,6 +1,6 @@
 package process.definitions;
 
 public class SquareRoot implements ArithmeticWithX1Only{
-    public int doOperation(int x1){ return (int) (Math.sqrt(x1));
+    public double doOperation(double x1){ return (int) (Math.sqrt(x1));
     }
 }
